Changed name of examples folder and refactored repository accordingly
