+++ .gitignore
@@ -1,4 +1,5 @@
 *.idea/
 *__pycache__/
-/examples/paper_results/gmm/models/
-/examples/many_well/models/
+*.ipynb_checkpoints/
+/experiments/paper_results/gmm/models/
+/experiments/many_well/models/
