--- examples/make_flow/__init__.py
@@ -1,4 +1,0 @@
-from examples.make_flow.make_realnvp_normflow import make_wrapped_normflowdist, make_normflow_model, make_normflow_snf_model
-from examples.make_flow.make_masked_affine_nflows import make_wrapped_nflows_dist
-
-# from examples.make_flow.make_aldp_model import make_aldp_model
