+++ experiments/gmm/README.md
@@ -3,22 +3,22 @@
 The following commands can each be used to train the methods from the paper:
 ```
 # FAB with prioritised buffer.
-python examples/gmm/run.py -m training.seed=0,1,2 training.use_buffer=True training.prioritised_buffer=True 
+python experiments/gmm/run.py -m training.seed=0,1,2 training.use_buffer=True training.prioritised_buffer=True 
 
 # FAB without the prioritised buffer.
-python examples/gmm/run.py -m training.seed=0,1,2 fab.loss_type=p2_over_q_alpha_2_div 
+python experiments/gmm/run.py -m training.seed=0,1,2 fab.loss_type=p2_over_q_alpha_2_div 
 
 # Flow using ground truth samples, training by maximum likelihood/forward KL divergence minimiation.
-python examples/gmm/run.py -m training.seed=0,1,2 fab.loss_type=target_forward_kl
+python experiments/gmm/run.py -m training.seed=0,1,2 fab.loss_type=target_forward_kl
 
 # Flow using alpha-divergence, with alpha=2
-python examples/gmm/run.py -m training.seed=0,1,2 fab.loss_type=flow_alpha_2_div_nis
+python experiments/gmm/run.py -m training.seed=0,1,2 fab.loss_type=flow_alpha_2_div_nis
 
 # Flow using reverse KL divergence
-python examples/gmm/run.py -m training.seed=0,1,2 fab.loss_type=flow_reverse_kld
+python experiments/gmm/run.py -m training.seed=0,1,2 fab.loss_type=flow_reverse_kld
 
 # SNF using reverse KLD
-python examples/gmm/run.py -m training.seed=0,1,2 flow.use_snf=True
+python experiments/gmm/run.py -m training.seed=0,1,2 flow.use_snf=True
 ```
 
 **Further notes** This will use hydra-multirun to run the random seeds in parallel. 
@@ -34,4 +34,4 @@ visualising each of the modes.
 
 ## Further Notes
 By 10 August 2022 we will upload a Colab notebook that installs this repository, runs 
-these examples and performs visualisation of results. 
+these experiments and performs visualisation of results.
