+++ experiments/gmm/plots/MoG.png
