+++ experiments/gmm/plots/MoG_appendix.png
