+++ experiments/many_well/results_vis_appendix.py
@@ -4,12 +4,12 @@ import matplotlib.pyplot as plt
 from matplotlib import rc
 import matplotlib as mpl
 from omegaconf import DictConfig
-from examples.make_flow import make_wrapped_normflowdist
-from examples.many_well.many_well_visualise_all_marginal_pairs import get_target_log_prob_marginal_pair
+from experiments.make_flow import make_wrapped_normflowdist
+from experiments.many_well.many_well_visualise_all_marginal_pairs import get_target_log_prob_marginal_pair
 from fab.utils.plotting import plot_contours, plot_marginal_pair
 from fab.target_distributions.many_well import ManyWellEnergy
 import torch
-from examples.setup_run_snf import make_normflow_snf_model, SNFModel
+from experiments.setup_run_snf import make_normflow_snf_model, SNFModel
 
 PATH = os.getcwd()
 
@@ -97,7 +97,7 @@ def run(cfg: DictConfig):
     for i, (ax, model_name, title) in enumerate(zip(subfigs[:len(titles)], model_names, titles)):
         plot_marginals(cfg, subfigs[i], model_names[i], plot_y_label=True)
         ax.suptitle(title)
-    plt.savefig("/home/laurence/work/code/FAB-TORCH/examples/paper_results/many_well/plots/many_well_appendix.png", bbox_inches="tight")
+    plt.savefig("/home/laurence/work/code/FAB-TORCH/experiments/paper_results/many_well/plots/many_well_appendix.png", bbox_inches="tight")
     plt.show()
